+++ django/applications/catmaid/static/js/widgets/stats.js
@@ -206,7 +206,7 @@
                         new THREE.Color().setRGB(1, 1, 0));
                     return o;
                   }, {});
-                  WindowMaker.create('neuron-staging-area', ST);
+                  WindowMaker.create('selection-table', ST);
                   ST.append(models);
                 }));
             break;